Changed return type of -pitch to float. Made both -setPitch: and -pitch use the fixed-point conversion macros in FixMath.h.

+++ Frameworks/Adium Framework/SUSpeaker.h
@@ -32,7 +32,7 @@
 -(void) speakText:(NSString*)text;
 -(void) stopSpeaking;
 -(void) resetToDefaults;
--(int) pitch;
+-(float) pitch;
 -(int) rate;
 -(NSString *)demoTextForVoiceAtIndex:(int)voiceIndex;
